services/carla-runner/requirements.txt
carla==0.9.15
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
pygame>=2.1.0
google-cloud-storage>=2.10.0
google-cloud-pubsub>=2.18.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0